refactor(employee): rename and update script paths for employee management
Renamed the `employee-script.js` and adjusted the corresponding paths in the project to reflect the new file location. This includes updating the import path in `index.js` and the script reference in the `employee.html` template. These changes improve the organization of the scripts and reduce the directory depth for easier access.
