+++ src/main/resources/templates/employee/employee.html
@@ -3,7 +3,7 @@
 <head>
     <title>Colaboradores</title>
     <link rel="stylesheet" th:href="@{protected/css/employee-styles.css}">
-    <script th:src="@{protected/js/employee/employee-script.js}" defer></script>
+    <script th:src="@{protected/js/employee-script.js}" defer></script>
 </head>
 <body>
 <main layout:fragment="content">
